Restructure repo: reduce root clutter from 29 to ~19 items
- Move test scripts to tests/ (test_discovery_local, test_p2p_e2e)
- Move cyberpunk.tcss to widgets/, update CSS_PATH in app.py
- Merge paths.py into config.py (platform-aware path definitions)
- Merge config_store.py into config.py (ConfigStore class)
- Move lid/engine.py to diarization/lid.py, remove lid/ package
- Move scripts/export_onnx.py to diarization/export_onnx.py, remove scripts/
- Move Dockerfile.test and requirements-dev.txt to dev/
- Update all imports across app.py, diagnostics.py, speakers/store.py,
  speakers/crypto.py, party/manager.py, diarization/engine.py,
  diarization/onnx_embedder.py, and tests/test_config_store.py
