+++ speakers/store.py
@@ -22,7 +22,7 @@ import numpy as np
 
 from .models import SpeakerMeta, SpeakerProfile
 from . import crypto
-from paths import DB_PATH as DEFAULT_DB_PATH, BACKUP_DIR
+from config import DB_PATH as DEFAULT_DB_PATH, BACKUP_DIR
 
 log = logging.getLogger(__name__)
