+++ tests/test_config_store.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from config_store import ConfigStore
+from config import ConfigStore
 
 
 @pytest.fixture
