+++ tests/test_discovery_local.py
@@ -14,7 +14,7 @@ import os
 import time
 import threading
 
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from network.discovery import PeerDiscovery
