+++ tests/test_p2p_e2e_standalone.py
@@ -18,7 +18,7 @@ import os
 import time
 import threading
 
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from network.discovery import PeerDiscovery
 from network.session import SessionManager
